Changed advancement trigger to only activate when player is currently in lava
