+++ src/main/java/by/dragonsurvivalteam/dragonsurvival/mixins/BlockMixin.java
@@ -18,8 +18,8 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public abstract class BlockMixin {
     @Inject(method="playerDestroy", at=@At(value="HEAD"))
     public void playerDestroy$dragonSurvival(Level pLevel, Player pPlayer, BlockPos pPos, BlockState pState, BlockEntity pBlockEntity, ItemStack pTool, CallbackInfo ci) {
-        if (pPlayer instanceof ServerPlayer serverPlayer) {
-            DSAdvancementTriggers.MINE_BLOCK.get().trigger(serverPlayer, (Block)(Object)(this));
+        if (pPlayer instanceof ServerPlayer serverPlayer && pPlayer.isInLava()) {
+            DSAdvancementTriggers.MINE_BLOCK_UNDER_LAVA.get().trigger(serverPlayer, (Block)(Object)(this));
         }
     }
 }
