+++ src/main/resources/assets/dragonsurvival/lang/en_us.json
@@ -1298,9 +1298,10 @@
   "ds.description.bolas": "Bolas (Crossbow): Transforms your arrows into bolas that slow enemies down.",
   "ds.advancements.be_dragon.title": "Be Yourself",
   "ds.advancements.be_dragon.description": "Become a dragon using an altar or any other method.",
-  "ds.advancements.be_root.title": "Elder Dragon Lore",
-  "ds.advancements.be_root.description": "Collect elder dragon dust by mining ore.",
-  "ds.advancements.be_root_2.title": "Dragon Starter Pack",
+  "ds.advancements.collect_dust.title": "Elder Dragon Lore",
+  "ds.advancements.collect_dust.description": "Collect elder dragon dust by mining ore.",
+  "ds.advancements.place_altar.title": "Dragon Starter Pack",
+  "ds.advancements.place_altar.description": "Place a Dragon Altar to change the species or appearance of a dragon.",
   "ds.advancements.fly_high.title": "Dragon Riders",
   "ds.advancements.fly_high.description": "Take a human friend high into the skies. Crouch!",
   "ds.advancements.sleep_on_treasure.title": "Loot: The Gathering",
@@ -1351,8 +1352,8 @@
   "ds.advancements.rain_dancing.description": "Feel hydration during rain and thunderstorm.",
   "ds.advancements.sea_peace_in_nether.title": "Heatproof",
   "ds.advancements.sea_peace_in_nether.description": "Use a Sea Beacon's magic to enter The Nether without drying out.",
-  "ds.advancements.sea_peace_in_nether.title": "Drought",
-  "ds.advancements.sea_peace_in_nether.description": "Place some snow to stay hydrated in the Nether.",
+  "ds.advancements.enter_snow_in_nether.title": "Drought",
+  "ds.advancements.enter_snow_in_nether.description": "Place some snow to stay hydrated in the Nether.",
   "ds.advancements.fish_eater.title": "No Rice in Sushi",
   "ds.advancements.fish_eater.description": "Eat all unique sea dragon foods from Dragon Survival. You can check what foods you can eat in the dragon altar.",
   "ds.advancements.master_all_sea_passives.title": "Hadopelagic Pressure",
@@ -1391,8 +1392,6 @@
   "ds.advancements.be_young_dragon.description": "As time goes on, the dragon grows...",
   "ds.advancements.be_adult_dragon.title": "Older And Stronger",
   "ds.advancements.be_adult_dragon.description": "Each day, you get stronger...",
-  "ds.advancements.be_full_growth_dragon.title": "Just Older",
-  "ds.advancements.be_full_growth_dragon.description": "Now you are old and strong, strong, strong!",
   "ds.advancements.collect_key.title": "Path Choice",
   "ds.advancements.collect_key.description": "Just exist.",
   "ds.mapstructures.hunters_castle": "Kingdom Explorer Map",
